# labelset: optimise String() function